+++ metaData.cc
@@ -301,7 +301,7 @@ void writeMetaData(struct metaData metadata, int globalFileNumber)
 		unsigned char input[strlen((char *)input1)];
 		strncpy((char *)metadata.fileID, (char *)input1, 20);
 //		strncpy((char *)input, (char *)input1+20, strlen((char *)input1)-20);
-		for(int i = 0 ; i < (strlen((char *)input1)-20) ; ++i)
+		for(int i = 0 ; i < (int)(strlen((char *)input1)-20) ; ++i)
 			input[i] = input1[i+20] ;
 
 		/*memset(metadata.fileID, '\0', sizeof(metadata.fileID));
@@ -394,7 +394,7 @@ void writeMetaData(struct metaData metadata, int globalFileNumber)
 		unsigned char input[input1.size()];
 		strncpy((char *)metadata.fileID, (char *)input1.c_str(), 20);
 //		strncpy((char *)input, (char *)input1+20, strlen((char *)input1)-20);
-		for(int i = 0 ; i < input1.size()-20 ; ++i)
+		for(int i = 0 ; i < (int)input1.size()-20 ; ++i)
 			input[i] = input1[i+20] ;
 
 		/*memset(metadata.fileID, '\0', sizeof(metadata.fileID));
